+++ python/sidecar/meta.py
@@ -1,23 +1,18 @@
+"""API for monitoring and controling the server"""
 import os
 
 import psutil
-from dotenv import load_dotenv
-from fastapi import FastAPI
+from fastapi import APIRouter
 
-load_dotenv()
-
-filesystem_api = FastAPI()  # API for interacting with the filesystem
-
-meta_api = FastAPI()
-"""API for monitoring and controling the server"""
+router = APIRouter()
 
 
-@meta_api.get("/status")
+@router.get("/status")
 async def status():
     return {"status": "ok"}
 
 
-@meta_api.post("/shutdown")
+@router.post("/shutdown")
 async def shutdown():
     # See https://stackoverflow.com/a/74757349/388951
     parent_pid = os.getpid()
