+++ python/sidecar/typing.py
@@ -36,43 +36,5 @@ class ResponseStatus(StrEnum):
 
 
 class CliCommands(RefStudioModel):
-    # Imports need to happen here to avoid circular imports
-    from sidecar.ai.schemas import (
-        ChatRequest,
-        ChatResponse,
-        RewriteRequest,
-        RewriteResponse,
-        TextCompletionRequest,
-        TextCompletionResponse,
-    )
-    from sidecar.references.schemas import (
-        DeleteRequest,
-        DeleteStatusResponse,
-        IngestRequest,
-        IngestResponse,
-        IngestStatusResponse,
-        ReferenceUpdate,
-        UpdateStatusResponse,
-    )
-    from sidecar.search.schemas import SearchRequest, SearchResponse
-
-    ingest: tuple[IngestRequest, IngestResponse]
-    """Ingest PDFs"""
-    ingest_status: tuple[None, IngestStatusResponse]
-    """Retrieve ingestion status of uploads"""
-    ingest_references: tuple[IngestRequest, IngestResponse]
-    """Retrieve ingested PDF references"""
-    rewrite: tuple[RewriteRequest, RewriteResponse]
-    """"Rewrites a block of text in a specified manner"""
-    completion: tuple[TextCompletionRequest, TextCompletionResponse]
-    """Completes a body of text"""
-    chat: tuple[ChatRequest, ChatResponse]
-    """Chat with the AI"""
-    update: tuple[ReferenceUpdate, UpdateStatusResponse]
-    """Update metadata for a Reference"""
-    delete: tuple[DeleteRequest, DeleteStatusResponse]
-    """Deletes a Reference"""
-    search: tuple[SearchRequest, SearchResponse]
-    """Searches for papers on Semantic Scholar"""
     serve: tuple[None, None]
     """Start an HTTP server"""
